Finish lab-12

+++ app/Http/Controllers/ShopController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Shop;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -28,6 +29,8 @@ class ShopController extends SearchableController
 
     function list(ServerRequestInterface $request): View
     {
+        Gate::authorize('list', Shop::class);
+
         $criteria = $this->prepareCriteria($request->getQueryParams());
         $query = $this->search($criteria)->withCount('products');
 
@@ -41,6 +44,8 @@ class ShopController extends SearchableController
     {
         $shop = $this->find($shopCode);
 
+        Gate::authorize('view', $shop);
+
         return view('shops.view', [
             'shop' => $shop,
         ]);
@@ -48,11 +53,15 @@ class ShopController extends SearchableController
 
     function showCreateForm(): View
     {
+        Gate::authorize('create', Shop::class);
+
         return view('shops.create-form');
     }
 
     function create(ServerRequestInterface $request): RedirectResponse
     {
+        Gate::authorize('create', Shop::class);
+
         $shop = Shop::create($request->getParsedBody());
 
         return redirect(
@@ -65,6 +74,8 @@ class ShopController extends SearchableController
     {
         $shop = $this->find($shopCode);
 
+        Gate::authorize('update', $shop);
+
         return view('shops.update-form', [
             'shop' => $shop,
         ]);
@@ -75,6 +86,9 @@ class ShopController extends SearchableController
         string $shopCode,
     ): RedirectResponse {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('update', $shop);
+
         $shop->fill($request->getParsedBody());
         $shop->save();
 
@@ -88,6 +102,9 @@ class ShopController extends SearchableController
     function delete(string $shopCode): RedirectResponse
     {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('delete', $shop);
+
         $shop->delete();
 
         return redirect(
@@ -102,6 +119,9 @@ class ShopController extends SearchableController
         string $shopCode,
     ): View {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('view', $shop);
+
         $criteria = $productController->prepareCriteria($request->getQueryParams());
         $query = $productController
             ->filter($shop->products(), $criteria)
@@ -121,6 +141,9 @@ class ShopController extends SearchableController
         string $shopCode,
     ): View {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('update', $shop);
+
         $query = $productController
             ->getQuery()
             ->whereDoesntHave(
@@ -149,6 +172,9 @@ class ShopController extends SearchableController
         string $shopCode,
     ): RedirectResponse {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('update', $shop);
+
         $data = $request->getParsedBody();
 
         $product = $productController
@@ -174,6 +200,9 @@ class ShopController extends SearchableController
         string $shopCode,
     ): RedirectResponse {
         $shop = $this->find($shopCode);
+
+        Gate::authorize('update', $shop);
+
         $data = $request->getParsedBody();
 
         $product = $shop->products()
